+++ Package.swift
@@ -48,8 +48,8 @@ let package = Package(
         ),
         .binaryTarget(
             name: "AttestationBindings",
-            url: "https://github.com/evervault/evervault-ios/releases/download/1.1.0/AttestationBindings.xcframework.zip",
-            checksum: "42c9b066eb75259f2e26c0f49d0736d7e4fb0e4b36c8d0f3643dda11957163d9"
+            url: "https://github.com/evervault/evervault-ios/releases/download/1.1.5/AttestationBindings.xcframework.zip",
+            checksum: "444c0cb0baab41754d6052994e482e8f096befa7b40eb4e7f77fdcb1d0e52bc7"
         ),
         .testTarget(
             name: "EvervaultCoreTests",