+++ Sources/EvervaultCore/core/DataType.swift
@@ -1,5 +1,5 @@
 import Foundation
 
 internal enum DataType: String {
-    case string, boolean, number
+    case string, boolean, number, file
 }