+++ Sources/EvervaultCore/core/dataHandlers/DataHandlers.swift
@@ -4,8 +4,8 @@ internal struct DataHandlers {
     struct Context: DataHandlerContext {
         let dataHandlers: DataHandlers
 
-        func encrypt(data: Any) throws -> Any {
-            try dataHandlers.encrypt(data: data)
+        func encrypt(data: Any, role: String?) throws -> Any {
+            try dataHandlers.encrypt(data: data, role: role)
         }
     }
 
@@ -22,13 +22,13 @@ internal struct DataHandlers {
         ]
     }
 
-    func encrypt(data: Any) throws -> Any {
+    func encrypt(data: Any, role: String?) throws -> Any {
         guard let handler = handlers.first(where: { $0.canEncrypt(data: data) }) else {
             throw CryptoError.notPossibleToHandleDataType
         }
         let context = Context(dataHandlers: self)
 
-        return try handler.encrypt(data: data, context: context)
+        return try handler.encrypt(data: data, role: role, context: context)
     }
 
 }